+++ cloudgraph-web/src/main/java/org/cloudgraph/web/model/configuration/PropertyEditBean.java
@@ -347,14 +347,14 @@ public class PropertyEditBean extends ModelBean {
     	return typeItems;
     }
     
-	public int getNameMaxLength() {
+	public long getNameMaxLength() {
 		PlasmaProperty nameProp = 
 			(PlasmaProperty)this.property.getType().getProperty(
 					Property.PTY_NAME);
 		return nameProp.getMaxLength();
 	}
 	
-	public int getDefinitionMaxLength() {
+	public long getDefinitionMaxLength() {
 		PlasmaProperty nameProp = 
 			(PlasmaProperty)this.property.getType().getProperty(
 					Property.PTY_DEFINITION);
@@ -372,7 +372,7 @@ public class PropertyEditBean extends ModelBean {
     	else
     		text = ((String)value).trim();
 
-    	int max = this.getDefinitionMaxLength();
+    	long max = this.getDefinitionMaxLength();
     	if (text.length() > max) {
             String msg = label + " is longer than allowed maximum "
                 + String.valueOf(max) + " characters";
