+++ cloudgraph-web/src/main/java/org/cloudgraph/web/model/data/EnumerationEditBean.java
@@ -204,7 +204,7 @@ public class EnumerationEditBean extends ModelBean {
 	    return result;
     }    
    
-	public int getNameMaxLength() {
+	public long getNameMaxLength() {
 		Type type = PlasmaTypeHelper.INSTANCE.getType(Classifier.class);
 		PlasmaProperty nameProp = 
 			(PlasmaProperty)type.getProperty(
@@ -212,7 +212,7 @@ public class EnumerationEditBean extends ModelBean {
 		return nameProp.getMaxLength();
 	}
 	
-	public int getDefinitionMaxLength() {
+	public long getDefinitionMaxLength() {
 		Type type = PlasmaTypeHelper.INSTANCE.getType(Classifier.class);
 		PlasmaProperty nameProp = 
 			(PlasmaProperty)type.getProperty(
@@ -220,7 +220,7 @@ public class EnumerationEditBean extends ModelBean {
 		return nameProp.getMaxLength();
 	}
 	
-	public int getLiteralNameMaxLength() {
+	public long getLiteralNameMaxLength() {
 		
 		Type type = PlasmaTypeHelper.INSTANCE.getType(EnumerationLiteral.class);
 		PlasmaProperty nameProp = 
@@ -239,7 +239,7 @@ public class EnumerationEditBean extends ModelBean {
     	else
     		text = ((String)value).trim();
 
-    	int max = this.getNameMaxLength();
+    	long max = this.getNameMaxLength();
     	if (text.length() > max) {
             String msg = label + " is longer than allowed maximum "
                 + String.valueOf(max) + " characters";
@@ -257,7 +257,7 @@ public class EnumerationEditBean extends ModelBean {
     	else
     		text = ((String)value).trim();
 
-    	int max = this.getDefinitionMaxLength();
+    	long max = this.getDefinitionMaxLength();
     	if (text.length() > max) {
             String msg = label + " is longer than allowed maximum "
                 + String.valueOf(max) + " characters";
@@ -275,7 +275,7 @@ public class EnumerationEditBean extends ModelBean {
     	else
     		text = ((String)value).trim();
 
-    	int max = this.getLiteralNameMaxLength();
+    	long max = this.getLiteralNameMaxLength();
     	if (text.length() > max) {
             String msg = label + " is longer than allowed maximum "
                 + String.valueOf(max) + " characters";
